Wizard CRUD done! Close #2

+++ src/main/java/com/omneya/hogwarts/hogwartsartifactsonline/repositories/WizardRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface WizardRepository extends JpaRepository<Wizard, Integer> {
+public interface WizardRepository extends JpaRepository<Wizard, Long> {
 }
